Extract file indexer path comparison partial
